+++ src/index.test.js
@@ -1,0 +1,9 @@
+const test = require("tape")
+
+test("index", t => {
+  t.doesNotThrow(() => {
+    require(".")
+  }, "Main file loads without errors")
+
+  t.end()
+})