+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.12.0] - 5 November 2018
+
+### Add
+
+- Add [`startsWith`](/src/starts-with/starts-with.js) - Test if string starts with substring
+
 ## [0.11.0] - 30 October 2018
 
 ### Add
@@ -87,8 +93,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Removed `flow` support
 
-[Unreleased]: https://github.com/codemachiner/m/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/codemachiner/m/compare/v0.12.0...HEAD
 
+[0.12.0]: https://github.com/codemachiner/m/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/codemachiner/m/compare/v0.10.1...v0.11.0
 [0.10.1]: https://github.com/codemachiner/m/compare/v0.10.0...v0.10.1
 [0.10.0]: https://github.com/codemachiner/m/compare/v0.9.0...v0.10.0